Update and rename backend_api.py to main_api.py
